+++ src/adapter/graphql/tests/tests/test_gql_dataset_flow_runs.rs
@@ -12,7 +12,7 @@
 use std::sync::Arc;
 
 use async_graphql::value;
-use chrono::Utc;
+use chrono::{DateTime, Duration, DurationRound, Utc};
 use container_runtime::ContainerRuntime;
 use dill::Component;
 use event_bus::EventBus;
@@ -33,6 +33,7 @@ use kamu_core::{
     DatasetRepository,
     DependencyGraphRepository,
     PollingIngestService,
+    PullResult,
     SystemTimeSourceDefault,
 };
 use kamu_flow_system::{
@@ -50,9 +51,9 @@ use kamu_flow_system_inmem::{
     FlowEventStoreInMem,
     FlowServiceInMemory,
 };
-use kamu_task_system::{TaskEventFinished, TaskEventRunning, TaskID, TaskOutcome};
+use kamu_task_system as ts;
 use kamu_task_system_inmem::{TaskSchedulerInMemory, TaskSystemEventStoreInMemory};
-use opendatafabric::{DatasetID, DatasetKind, FAKE_ACCOUNT_ID};
+use opendatafabric::{DatasetID, DatasetKind, Multihash, FAKE_ACCOUNT_ID};
 
 use crate::utils::{authentication_catalogs, expect_anonymous_access_error};
 
@@ -60,7 +61,12 @@ use crate::utils::{authentication_catalogs, expect_anonymous_access_error};
 
 #[test_log::test(tokio::test)]
 async fn test_trigger_ingest_root_dataset() {
-    let harness = FlowRunsHarness::new();
+    let mut dependency_graph_repo_mock = MockDependencyGraphRepository::new();
+    dependency_graph_repo_mock
+        .expect_list_dependencies_of_all_datasets()
+        .return_once(|| Box::pin(futures::stream::empty()));
+
+    let harness = FlowRunsHarness::new_custom(dependency_graph_repo_mock);
     let create_result = harness.create_root_dataset().await;
 
     let mutation_code =
@@ -122,7 +128,7 @@ async fn test_trigger_ingest_root_dataset() {
                                         "description": {
                                             "__typename": "FlowDescriptionDatasetPollingIngest",
                                             "datasetId": create_result.dataset_handle.id.to_string(),
-                                            "ingestedRecordsCount": null,
+                                            "ingestResult": null,
                                         },
                                         "status": "QUEUED",
                                         "outcome": null,
@@ -158,13 +164,235 @@ async fn test_trigger_ingest_root_dataset() {
             }
         })
     );
+
+    let flow_task_id = harness.mimic_flow_scheduled("0").await;
+
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{response:?}");
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetPollingIngest",
+                                            "datasetId": create_result.dataset_handle.id.to_string(),
+                                            "ingestResult": null,
+                                        },
+                                        "status": "SCHEDULED",
+                                        "outcome": null,
+                                        "timing": {
+                                            "runningSince": null,
+                                            "finishedAt": null,
+                                        },
+                                        "tasks": [
+                                            {
+                                                "taskId": "0",
+                                                "status": "QUEUED",
+                                                "outcome": null,
+                                            }
+                                        ],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
+
+    let running_time = Utc::now().duration_round(Duration::seconds(1)).unwrap();
+    harness.mimic_task_running(flow_task_id, running_time).await;
+
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{response:?}");
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetPollingIngest",
+                                            "datasetId": create_result.dataset_handle.id.to_string(),
+                                            "ingestResult": null,
+                                        },
+                                        "status": "RUNNING",
+                                        "outcome": null,
+                                        "timing": {
+                                            "runningSince": running_time.to_rfc3339(),
+                                            "finishedAt": null,
+                                        },
+                                        "tasks": [
+                                            {
+                                                "taskId": "0",
+                                                "status": "RUNNING",
+                                                "outcome": null,
+                                            }
+                                        ],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
+
+    let complete_time = Utc::now().duration_round(Duration::seconds(1)).unwrap();
+    harness
+        .mimic_task_completed(
+            flow_task_id,
+            complete_time,
+            ts::TaskOutcome::Success(ts::TaskResult::PullResult(PullResult::Updated {
+                old_head: Some(Multihash::from_digest_sha3_256(b"old-slice")),
+                new_head: Multihash::from_digest_sha3_256(b"new-slice"),
+                num_blocks: 1,
+                num_records: 12,
+            })),
+        )
+        .await;
+
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{response:?}");
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetPollingIngest",
+                                            "datasetId": create_result.dataset_handle.id.to_string(),
+                                            "ingestResult": {
+                                                "numBlocks": 1,
+                                                "numRecords": 12,
+                                            },
+                                        },
+                                        "status": "FINISHED",
+                                        "outcome": "SUCCESS",
+                                        "timing": {
+                                            "runningSince": running_time.to_rfc3339(),
+                                            "finishedAt": complete_time.to_rfc3339(),
+                                        },
+                                        "tasks": [
+                                            {
+                                                "taskId": "0",
+                                                "status": "FINISHED",
+                                                "outcome": "SUCCESS",
+                                            }
+                                        ],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////
 
 #[test_log::test(tokio::test)]
 async fn test_trigger_execute_transform_derived_dataset() {
-    let harness = FlowRunsHarness::new();
+    let mut dependency_graph_repo_mock = MockDependencyGraphRepository::new();
+    dependency_graph_repo_mock
+        .expect_list_dependencies_of_all_datasets()
+        .return_once(|| Box::pin(futures::stream::empty()));
+
+    let harness = FlowRunsHarness::new_custom(dependency_graph_repo_mock);
     harness.create_root_dataset().await;
     let create_derived_result = harness.create_derived_dataset().await;
 
@@ -229,7 +457,7 @@ async fn test_trigger_execute_transform_derived_dataset() {
                                         "description": {
                                             "__typename": "FlowDescriptionDatasetExecuteTransform",
                                             "datasetId": create_derived_result.dataset_handle.id.to_string(),
-                                            "transformedRecordsCount": null,
+                                            "transformResult": null,
                                         },
                                         "status": "QUEUED",
                                         "outcome": null,
@@ -238,6 +466,93 @@ async fn test_trigger_execute_transform_derived_dataset() {
                                             "finishedAt": null,
                                         },
                                         "tasks": [],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
+
+    let flow_task_id = harness.mimic_flow_scheduled("0").await;
+
+    let running_time = Utc::now().duration_round(Duration::seconds(1)).unwrap();
+    harness.mimic_task_running(flow_task_id, running_time).await;
+
+    let complete_time = Utc::now().duration_round(Duration::seconds(1)).unwrap();
+    harness
+        .mimic_task_completed(
+            flow_task_id,
+            complete_time,
+            ts::TaskOutcome::Success(ts::TaskResult::PullResult(PullResult::Updated {
+                old_head: Some(Multihash::from_digest_sha3_256(b"old-slice")),
+                new_head: Multihash::from_digest_sha3_256(b"new-slice"),
+                num_blocks: 3,
+                num_records: 25,
+            })),
+        )
+        .await;
+
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{response:?}");
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetExecuteTransform",
+                                            "datasetId": create_derived_result.dataset_handle.id.to_string(),
+                                            "transformResult": {
+                                                "numBlocks": 3,
+                                                "numRecords": 25,
+                                            },
+                                        },
+                                        "status": "FINISHED",
+                                        "outcome": "SUCCESS",
+                                        "timing": {
+                                            "runningSince": running_time.to_rfc3339(),
+                                            "finishedAt": complete_time.to_rfc3339(),
+                                        },
+                                        "tasks": [
+                                            {
+                                                "taskId": "0",
+                                                "status": "FINISHED",
+                                                "outcome": "SUCCESS",
+                                            }
+                                        ],
                                         "initiator": {
                                             "id": FAKE_ACCOUNT_ID,
                                             "accountName": auth::DEFAULT_ACCOUNT_NAME,
@@ -1143,8 +1458,14 @@ async fn test_cancel_already_succeeded_flow() {
     let flow_id = FlowRunsHarness::extract_flow_id_from_trigger_response(&response_json);
 
     let flow_task_id = harness.mimic_flow_scheduled(flow_id).await;
-    harness.mimic_task_running(flow_task_id).await;
-    harness.mimic_task_completed(flow_task_id).await;
+    harness.mimic_task_running(flow_task_id, Utc::now()).await;
+    harness
+        .mimic_task_completed(
+            flow_task_id,
+            Utc::now(),
+            ts::TaskOutcome::Success(ts::TaskResult::Empty),
+        )
+        .await;
 
     let mutation_code =
         FlowRunsHarness::cancel_scheduled_tasks_mutation(&create_result.dataset_handle.id, flow_id);
@@ -1213,8 +1534,14 @@ async fn test_history_of_completed_flow() {
         .await;
 
     let flow_task_id = harness.mimic_flow_scheduled(flow_id).await;
-    harness.mimic_task_running(flow_task_id).await;
-    harness.mimic_task_completed(flow_task_id).await;
+    harness.mimic_task_running(flow_task_id, Utc::now()).await;
+    harness
+        .mimic_task_completed(
+            flow_task_id,
+            Utc::now(),
+            ts::TaskOutcome::Success(ts::TaskResult::Empty),
+        )
+        .await;
 
     let query = FlowRunsHarness::flow_history_query(&create_result.dataset_handle.id, flow_id);
 
@@ -1367,8 +1694,8 @@ impl FlowRunsHarness {
             .add::<FlowServiceInMemory>()
             .add::<FlowEventStoreInMem>()
             .add_value(FlowServiceRunConfig::new(
-                chrono::Duration::seconds(1),
-                chrono::Duration::minutes(1),
+                Duration::seconds(1),
+                Duration::minutes(1),
             ))
             .add::<TaskSchedulerInMemory>()
             .add::<TaskSystemEventStoreInMemory>()
@@ -1430,7 +1757,7 @@ impl FlowRunsHarness {
             .unwrap()
     }
 
-    async fn mimic_flow_scheduled(&self, flow_id: &str) -> TaskID {
+    async fn mimic_flow_scheduled(&self, flow_id: &str) -> ts::TaskID {
         let flow_service_test_driver = self
             .catalog_authorized
             .get_one::<dyn FlowServiceTestDriver>()
@@ -1460,36 +1787,63 @@ impl FlowRunsHarness {
         flow.save(flow_event_store.as_ref()).await.unwrap();
     }
 
-    async fn mimic_task_running(&self, task_id: TaskID) {
+    async fn mimic_task_running(&self, task_id: ts::TaskID, event_time: DateTime<Utc>) {
         let flow_service_test_driver = self
             .catalog_authorized
             .get_one::<dyn FlowServiceTestDriver>()
             .unwrap();
         flow_service_test_driver.mimic_running_started();
 
+        let task_event_store = self
+            .catalog_anonymous
+            .get_one::<dyn ts::TaskSystemEventStore>()
+            .unwrap();
+
+        let mut task = ts::Task::load(task_id, task_event_store.as_ref())
+            .await
+            .unwrap();
+        task.run(event_time).unwrap();
+        task.save(task_event_store.as_ref()).await.unwrap();
+
         let event_bus = self.catalog_authorized.get_one::<EventBus>().unwrap();
         event_bus
-            .dispatch_event(TaskEventRunning {
-                event_time: Utc::now(),
+            .dispatch_event(ts::TaskEventRunning {
+                event_time,
                 task_id,
             })
             .await
             .unwrap();
     }
 
-    async fn mimic_task_completed(&self, task_id: TaskID) {
+    async fn mimic_task_completed(
+        &self,
+        task_id: ts::TaskID,
+        event_time: DateTime<Utc>,
+        task_outcome: ts::TaskOutcome,
+    ) {
         let flow_service_test_driver = self
             .catalog_authorized
             .get_one::<dyn FlowServiceTestDriver>()
             .unwrap();
         flow_service_test_driver.mimic_running_started();
 
+        let task_event_store = self
+            .catalog_anonymous
+            .get_one::<dyn ts::TaskSystemEventStore>()
+            .unwrap();
+
+        let mut task = ts::Task::load(task_id, task_event_store.as_ref())
+            .await
+            .unwrap();
+        task.finish(event_time, task_outcome.clone()).unwrap();
+        task.save(task_event_store.as_ref()).await.unwrap();
+
         let event_bus = self.catalog_authorized.get_one::<EventBus>().unwrap();
         event_bus
-            .dispatch_event(TaskEventFinished {
-                event_time: Utc::now(),
+            .dispatch_event(ts::TaskEventFinished {
+                event_time,
                 task_id,
-                outcome: TaskOutcome::Success,
+                outcome: task_outcome,
             })
             .await
             .unwrap();
@@ -1521,17 +1875,26 @@ impl FlowRunsHarness {
                                             }
                                             ... on FlowDescriptionDatasetExecuteTransform {
                                                 datasetId
-                                                transformedRecordsCount
+                                                transformResult {
+                                                    numBlocks
+                                                    numRecords
+                                                }
                                             }
                                             ... on FlowDescriptionDatasetPollingIngest {
                                                 datasetId
-                                                ingestedRecordsCount
+                                                ingestResult {
+                                                    numBlocks
+                                                    numRecords
+                                                }
                                             }
                                             ... on FlowDescriptionDatasetPushIngest {
                                                 datasetId
                                                 sourceName
                                                 inputRecordsCount
-                                                ingestedRecordsCount
+                                                ingestResult {
+                                                    numBlocks
+                                                    numRecords
+                                                }
                                             }
                                         }
                                         status