+++ src/infra/core/tests/tests/test_pull_service_impl.rs
@@ -1230,6 +1230,7 @@ impl SyncService for TestSyncService {
                     old_head: None,
                     new_head: Multihash::from_digest_sha3_256(b"boop"),
                     num_blocks: 1,
+                    num_records: 0,
                 }),
             });
         }