+++ src/infra/task-system-inmem/src/task_executor_inmem.rs
@@ -87,12 +87,14 @@ impl TaskExecutor for TaskExecutorInMemory {
             let outcome = match &task.logical_plan {
                 LogicalPlan::UpdateDataset(upd) => {
                     let pull_svc = self.catalog.get_one::<dyn PullService>().int_err()?;
-                    let res = pull_svc
+                    let maybe_pull_result = pull_svc
                         .pull(&upd.dataset_id.as_any_ref(), PullOptions::default(), None)
                         .await;
 
-                    match res {
-                        Ok(_) => TaskOutcome::Success,
+                    match maybe_pull_result {
+                        Ok(pull_result) => {
+                            TaskOutcome::Success(TaskResult::PullResult(pull_result))
+                        }
                         Err(_) => TaskOutcome::Failed,
                     }
                 }
@@ -104,7 +106,9 @@ impl TaskExecutor for TaskExecutorInMemory {
                     if let Some(busy_time) = busy_time {
                         tokio::time::sleep(*busy_time).await;
                     }
-                    end_with_outcome.unwrap_or(TaskOutcome::Success)
+                    end_with_outcome
+                        .clone()
+                        .unwrap_or(TaskOutcome::Success(TaskResult::Empty))
                 }
             };
 
@@ -117,7 +121,8 @@ impl TaskExecutor for TaskExecutorInMemory {
 
             // Refresh the task in case it was updated concurrently (e.g. late cancellation)
             task.update(self.event_store.as_ref()).await.int_err()?;
-            task.finish(self.time_source.now(), outcome).int_err()?;
+            task.finish(self.time_source.now(), outcome.clone())
+                .int_err()?;
             task.save(self.event_store.as_ref()).await.int_err()?;
 
             self.publish_task_finished(task.task_id, outcome).await?;